Moved the UpgradePrompt to the v3 folder
